Revert "rsx/vk: Implement hardware instancing (#16466)"
This reverts commit 62701154f1a6ebd71499cabba2341cc240ff50c4.
