+++ rpcs3/Emu/RSX/Program/GLSLTypes.h
@@ -22,7 +22,6 @@ namespace glsl
 		// Applicable in vertex stage
 		bool require_lit_emulation : 1;
 		bool require_explicit_invariance : 1;
-		bool require_instanced_render : 1;
 		bool emulate_zclip_transform : 1;
 		bool emulate_depth_clip_only : 1;
