+++ rpcs3/Emu/RSX/Program/VertexProgramDecompiler.h
@@ -132,10 +132,6 @@ protected:
 public:
 	struct
 	{
-		// Configuration properties (in)
-		// None
-
-		// Decoded properties (out)
 		bool has_lit_op = false;
 		bool has_indexed_constants = false;
 	}
