+++ rpcs3/Emu/RSX/VK/VKGSRenderTypes.hpp
@@ -197,7 +197,6 @@ namespace vk
 		s64 index_heap_ptr = 0;
 		s64 texture_upload_heap_ptr = 0;
 		s64 rasterizer_env_heap_ptr = 0;
-		s64 instancing_heap_ptr = 0;
 
 		u64 last_frame_sync_time = 0;
 
@@ -219,7 +218,6 @@ namespace vk
 			index_heap_ptr = other.index_heap_ptr;
 			texture_upload_heap_ptr = other.texture_upload_heap_ptr;
 			rasterizer_env_heap_ptr = other.rasterizer_env_heap_ptr;
-			instancing_heap_ptr = other.instancing_heap_ptr;
 		}
 
 		// Exchange storage (non-copyable)
@@ -231,7 +229,7 @@ namespace vk
 		void tag_frame_end(
 			s64 attrib_loc, s64 vtxenv_loc, s64 fragenv_loc, s64 vtxlayout_loc,
 			s64 fragtex_loc, s64 fragconst_loc, s64 vtxconst_loc, s64 index_loc,
-			s64 texture_loc, s64 rasterizer_loc, s64 instancing_loc)
+			s64 texture_loc, s64 rasterizer_loc)
 		{
 			attrib_heap_ptr = attrib_loc;
 			vtx_env_heap_ptr = vtxenv_loc;
@@ -243,7 +241,6 @@ namespace vk
 			index_heap_ptr = index_loc;
 			texture_upload_heap_ptr = texture_loc;
 			rasterizer_env_heap_ptr = rasterizer_loc;
-			instancing_heap_ptr = instancing_loc;
 
 			last_frame_sync_time = rsx::get_shared_tag();
 		}
