+++ rpcs3/Emu/RSX/VK/vkutils/pipeline_binding_table.h
@@ -14,10 +14,8 @@ namespace vk
 		u8 vertex_buffers_first_bind_slot      = 5;
 		u8 conditional_render_predicate_slot   = 8;
 		u8 rasterizer_env_bind_slot            = 9;
-		u8 instancing_lookup_table_bind_slot   = 10;
-		u8 instancing_constants_buffer_slot    = 11;
-		u8 textures_first_bind_slot            = 12;
-		u8 vertex_textures_first_bind_slot     = 12;                              // Invalid, has to be initialized properly
+		u8 textures_first_bind_slot            = 10;
+		u8 vertex_textures_first_bind_slot     = 10;                              // Invalid, has to be initialized properly
 		u8 total_descriptor_bindings           = vertex_textures_first_bind_slot; // Invalid, has to be initialized properly
 	};
 }
