+++ rpcs3/emucore.vcxproj.filters
@@ -1315,9 +1315,6 @@
     <ClCompile Include="Emu\RSX\Host\MM.cpp">
       <Filter>Emu\GPU\RSX\Host Mini-Driver</Filter>
     </ClCompile>
-    <ClCompile Include="Emu\RSX\Core\RSXDrawCommands.cpp">
-      <Filter>Emu\GPU\RSX\Core</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Crypto\aes.h">
@@ -2652,12 +2649,6 @@
     </ClInclude>
     <ClInclude Include="Emu\RSX\Host\MM.h">
       <Filter>Emu\GPU\RSX\Host Mini-Driver</Filter>
-    </ClInclude>
-    <ClInclude Include="Emu\RSX\Core\RSXDrawCommands.h">
-      <Filter>Emu\GPU\RSX\Core</Filter>
-    </ClInclude>
-    <ClInclude Include="Emu\RSX\Core\RSXDriverState.h">
-      <Filter>Emu\GPU\RSX\Core</Filter>
     </ClInclude>
     <ClInclude Include="Emu\NP\fb_helpers.h">
       <Filter>Emu\NP</Filter>
